+++ changelog.md
@@ -1,3 +1,8 @@
+# 0.0.6
+
+* Prevent gateways outside the world border
+* Added structurify compat
+
 # 0.0.5
 
 * Fixed issues with config loading